Finish putting methods in alg
