+++ src/pyfoma/__init__.py
@@ -1,4 +1,9 @@
-from pyfoma.fst import FST
+from pyfoma.fst import *
+from pyfoma.fst import re as regex
+from pyfoma.paradigm import Paradigm
+from pyfoma.algorithms import concatenate, union, intersection, kleene_star, kleene_plus, difference, cross_product, compose, optional, \
+                                ignore, projected, inverted, reversed, reversed_e, minimized, minimized_as_dfa, determinized_as_dfa, determinized_unweighted
+
 
 __author__     = "Mans Hulden"
 __copyright__  = "Copyright 2022"
